requirements.txt
click==6.6
ecdsa==0.13
Fabric3==1.12.post1
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
paramiko==1.17.2
pycrypto==2.6.1
six==1.10.0
Werkzeug==0.11.11
